Optimized MT
Cracker will now calculate hashes on relatively "small" chunks of passwords and check if any of the hashes matches the shadow file.  Also dispenses with hash map.  Significantly reduces the amount of memory that needs to be committed. Can be left unattended and will process "rockyou" in about 2 hours with an aggressive threading scheme.  A less aggressive scheme is used in this commit to not bog down the rest of the system.